# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: "API for managing user registration, authentication, and profiles."
  title: User Management API
  version: 1.0.0
servers:
- url: /
paths:
  /users:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/User"
                type: array
          description: A list of users
      summary: Get all users
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          description: User created successfully
      summary: Create a new user
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/{userId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: User deleted successfully
      summary: Delete a user by ID
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: A single user
      summary: Get a user by ID
      x-accepts:
      - application/json
    put:
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: User updated successfully
      summary: Update a user by ID
      x-content-type: application/json
      x-accepts:
      - application/json
components:
  schemas:
    User:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        user_id: 0
        password_hash: password_hash
        created_at: 2000-01-23T04:56:07.000+00:00
        email: email
        username: username
      properties:
        user_id:
          type: integer
        username:
          type: string
        email:
          type: string
        password_hash:
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      type: object